Create Background_Remove(using API).py
This code uses FastAPI application  which provides an endpoint to process images. It downloads an image from a provided URL, optionally crops it based on a specified bounding box, removes the background using the rembg library, and uploads the processed image to Google Cloud Storage. The application then returns a public URL to access the processed image.